requirements.txt
transformers>=4.51.0
torch>=2.0.0
ollama>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
chromadb>=0.4.0
pandas>=2.0.0
dashscope>=1.14.0
python-dotenv>=1.0.0
jinja2>=3.1.0
lightrag>=1.0.0 